Enhance the CFI pass to generate unique CFI labels.
Enhance the forward-edge CFI pass to generate globally unique CFI
labels.  The new CFI label (0xf870f871) is chosen to be a concatenation
of two 16-bit undefined encodings in Thumb/Thumb-2 instruction set, in
which way we can guarantee that such a byte sequence will not be
generated by the compiler except for our CFI label.

The placement of the CFI label is also updated to use a Function's
prefix data; this eliminates the need for the CFI label to be a NOP
because it will not be jumped to and executed.