Lower to runtime libcalls of different names.
Modify the code generator to lower IR instructions that end up being
runtime libcalls to ones with a different name based on whether the
containing function is in a privileged code section.  For example, a
double-precision floating-point division will be lowered to a call to
`__aeabi_ddivPriv()` instead of `__aeabi_ddiv()` if the function is in
the `privileged_functions` section.

This is achieved by the following changes:

* Intercept calls to `TargetLowering::setLibcallName()` and
  `TargetLowering::getLibcallName()` by making the two methods virtual
  and overriding them in `ARMTargetLowering`, in which a separate table
  of names for privileged runtime libcalls is maintained.

* Create different instances of `ARMSubtarget` by attaching a ",Priv"
  suffix at the end of each key for getting the `ARMSubtarget` of a
  privileged function.

* Add a field and a method in `ARMSubtarget` to indicate whether this
  instance is for privileged functions, and use this information in
  `ARMTargetLowering::getLibcallName()` to return the appropriate name.